Split the main exco.py file into modules
