+++ exco/extensions.py
@@ -1,0 +1,5 @@
+from flask_sqlalchemy import SQLAlchemy
+
+# define DB for import in other module
+db = SQLAlchemy()
+
